requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
openai==1.74.0
pydantic==2.6.3
PyPDF2==3.0.1
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
sentence-transformers==4.1.0
tiktoken==0.9.0
torch==2.6.0
transformers==4.51.3
python-dotenv
pyyaml
html2text
